Crate rust_enum_derive

Structs

FileArgs
Arguments for how to process() an input file.

Functions

process
This is the function that you call to process one file (Enum) worth of data.

traverse_dir
This is the function that you call to process a whole directory heirarcy full of files.